#2158: relocate wg/engine.rs inline tests to engine_tests.rs sibling
Pure code-motion file-split (1 of N for #2158), the #1046 frame/tests.rs
pattern. engine.rs had crossed the ~2000 prod-LOC modularity threshold
(2086 audit-LOC) almost entirely on the back of its ~809-line inline
#[cfg(test)] mod engine_internal_tests block.

Cut that whole test module out of engine.rs into a new sibling
userspace-dp/src/afxdp/wg/engine_tests.rs and re-declare it in engine.rs
with:

    #[cfg(test)]
    #[path = "engine_tests.rs"]
    mod engine_internal_tests;

Because #[path] keeps the module a child of wg::engine, the test block's
`use super::*;` (and the rest of its imports) resolve exactly as before,
so no test body changed. The relocated bodies are byte-identical to the
former inline module modulo a uniform 4-space dedent; engine.rs's
production code (lines 1-1277) is byte-identical to master. No production
fn, type, or visibility was touched. The two #[cfg(test)] prod-file test
hooks (mock_now_ns field, table_for_test accessor) stay in engine.rs.

engine.rs drops 2086 -> 1280 LOC, off the REFACTOR/WATCH list entirely.
engine_tests.rs is test-named so the audit tool excludes it from the
heatmap.

Validation:
  - make build-userspace-dp equivalent release build: clean (exit 0).
  - cargo test --release wg::: 145 passed / 0 failed (same count and same
    tests as master; the relocated module now runs under
    afxdp::wg::engine::engine_internal_tests::*, an unchanged module path).
  - wg #[test] census unchanged vs origin/master (engine.rs 14 -> 0,
    engine_tests.rs 0 -> 14; all other wg files unchanged).
